% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codoncount.R
\name{aggregateCodonCount}
\alias{aggregateCodonCount}
\title{aggregateCodonCount}
\usage{
aggregateCodonCount(strings_codon, group_table)
}
\arguments{
\item{strings_codon}{the codon count table generated by getCodonCount(strings)}

\item{group}{an additional table for grouping information of sequences}
}
\description{
aggregate the codon counts for a given set of grouping for the initial set of sequences
}
