% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{getCDSfromGTF}
\alias{getCDSfromGTF}
\title{getCDSfromGTF}
\usage{
getCDSfromGTF(gtf, seqs, attrsep = "; ")
}
\arguments{
\item{gtf}{the path to the gtf file}

\item{seqs}{the sequences that are associated with the gtf file}

\item{attrsep}{in the attributes column of gtf, how are attributes separated? Default "; ".}
}
\description{
Get the full spliced coding sequences for each transcript
}
