% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonbias.R
\name{getCodonBias}
\alias{getCodonBias}
\title{getCodonBias}
\usage{
getCodonBias(counts, method = "RSCU", sizeFactor = 100)
}
\arguments{
\item{counts}{counts for each codon}

\item{method}{method of codon pair bias calculation, e.g. RSCU}

\item{sizeFactor}{the size factor when size normalization is used}
}
\description{
Generate Codon bias measures for each codon and sequence
}
