% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codoncount.R
\name{getCodonCount}
\alias{getCodonCount}
\title{getCodonCount}
\usage{
getCodonCount(strings, mode = "singleton", string_chunks = 20)
}
\arguments{
\item{strings}{a set of sequences to count codons}

\item{mode}{codon counting mode: singleton or pair}

\item{string_chunks}{the number of chunks to divide sequence list, default: 20}
}
\description{
get singleton or pair codon count table for a set of sequences
}
