% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonbias.R
\name{getCodonPairBias}
\alias{getCodonPairBias}
\title{getCodonPairBias}
\usage{
getCodonPairBias(
  pair_counts,
  single_counts,
  method = "RSCPU",
  GENETIC_CODE_PAIR
)
}
\arguments{
\item{pair_counts}{counts for each codon pair}

\item{single_counts}{counts for each codon}

\item{method}{method of codon pair bias calculation, e.g. RSCPU}

\item{GENETIC_CODE_PAIR}{}
}
\description{
Generate Codon pair bias measures for each codon pair and sequence
}
